Match video + custom controls functionality to react-player's example App
